backend/requirements.txt
# Django and Web Framework
Django==4.2.23
djangorestframework==3.16.0
django-cors-headers==4.7.0
asgiref==3.9.1

# Database
psycopg2-binary==2.9.10
dj-database-url==3.0.1
sqlparse==0.5.3

# Authentication and Security
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0

# Data Processing and Analysis
numpy==1.24.3
pandas==2.3.1

# Email
resend==0.8.0

# Utilities
python-dotenv==1.1.1
packaging==25.0
typing_extensions==4.14.1

# Production Server
gunicorn==23.0.0

# Development and Testing (optional)
# pytest==7.4.3
# pytest-django==4.7.0
# black==23.11.0
# flake8==6.1.0
